refactor: updated `SocijalniNatjecajBodovi.UkupnoBodova` column type from `int` to `float` for improved precision and compatibility

+++ Models/SocijalniNatjecajBodovi.cs
@@ -30,5 +30,5 @@ public class SocijalniNatjecajBodovi : AuditableEntity
     public float BodoviObrana { get; set; }
     public byte BodoviSeksualnoNasilje { get; set; }
     public byte BodoviCivilniStradalnici { get; set; }
-    public ushort UkupnoBodova { get; set; }
+    public float UkupnoBodova { get; set; }
 }
